pyproject.toml
[project]
name = "telusko"
version = "0.1.0"
description = "Telusko AI engg course Notes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "fastmcp>=2.14.2",
    "google-genai>=1.51.0",
    "langchain>=1.2.2",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
